requirements.txt
streamlit>=1.51.0
streamlit-pdf-viewer>=0.0.26
openai>=1.0.0
google-generativeai>=0.3.0
scikit-learn>=1.0.0

